:recycle: post와 user API 분리
