+++ src/entities/post/api/index.ts
@@ -1,0 +1,7 @@
+import { PostResponse } from "../types"
+
+export const fetchPosts = async (limit: number, skip: number) => {
+  const response = await fetch(`/api/posts?limit=${limit}&skip=${skip}`)
+  const data: PostResponse = await response.json()
+  return data
+}
